+++ dnsgenerator.h
@@ -53,7 +53,7 @@ struct _DnsGenerator {
                        const char *Source,
                        BOOL IncludeRecordCounts
                        );
-    void (*CopyIdentifier)(DnsGenerator *g, int Value);
+    void (*CopyIdentifier)(DnsGenerator *g, uint16_t Value);
     int (*CopyCName)(DnsGenerator *g, DnsSimpleParserIterator *i);
     int (*CopyA)(DnsGenerator *g, DnsSimpleParserIterator *i);
     int (*CopyAAAA)(DnsGenerator *g, DnsSimpleParserIterator *i);
